@@ -52,10 +52,10 @@
         "joomla/session": "^3.0",
         "joomla/test": "^3.0",
         "joomla/uri": "^3.0",
-        "phpunit/phpunit": "^9.5.28",
-        "symfony/phpunit-bridge": "^5.0",
-        "squizlabs/php_codesniffer": "~3.7.1",
-        "rector/rector": "^0.15.21",
+        "phpunit/phpunit": "^10.0",
+        "symfony/phpunit-bridge": "^7.0",
+        "squizlabs/php_codesniffer": "~3.10.2",
+        "rector/rector": "^1.0",
         "phpstan/phpstan": "^1.10.7",
         "phan/phan": "^5.4"
     },
